test-freeze: incomplete test of prototype checking
The goal is to test that we haven't made a specific mistake in writing
harden(). In the correct implementation, if we throw an error because a
prototype of something we're hardening won't be frozen by the time we're
done (because it wasn't already frozen, and it wasn't something we visited
during harden()), then we don't mark *anything* as hardened: we add nothing
to the is-frozen set. In the errorful implementation, we add everything we
freeze to the is-frozen set before checking the prototypes and throwing
errors.

The case that's different is when C inherits from B, B inherits from A, none
are previously frozen, we harden B (which errors because A wasn't frozen),
then we attempt to harden C. The correct implementation should error when we
try to harden C, because B is not marked as hardened. The errorful
implementation accepts C because it thinks B *is* marked as hardened.

The key to clarity is the difference between being frozen (which is shallow)
and being hardened (which is transitive). You can only get into the hardened
set if all your prototypes are already in (or will soon be added to) the
hardened set. The errorful implementation *freezes* some things, but that
isn't enough to let them qualify as hardened.